requirements.txt
streamlit
numpy
pandas
scikit-learn
faiss-cpu
sentence-transformers
torch
requests
